metasys: support some newer Legrand devices and fix a few issues
Add to the metasys driver a few new models (now branded Legrand) speaking the Meta System UPS protocol.

Also:
- treat any '\0' in the serial number as a '0',
- for bypass mode, call status_set() with "BYPASS", not "BY",
- retrieve also battery charge and runtime,
- update support level of Meta System devices listed in the provided protocol: see nut-website's 570c1e5e9f1fb4f82abef9359897b5159e88fc50

Lastly, as per protocol documentation, data is sent by the device with unsigned integers of 8, 16 or 32 bits, with errors/special values being:
- for uint16: 0xFFFF and 0xFFFE,
- for uint32: 0xFFFFFFFF.
Alas, those values are (and, probably, have always been) actually reported in the documentation (mostly) as if their types were not unsigned (so, in decimal: -1, and -2) and the metasys driver used them like that, converting the values to non-fixed width signed types (int and long) and then expecting a value of -1 or -2 to signal a particular event, but this only happens on platforms where int is exactly 16 bits, and long 32 bits.
So, leave values as unsigned types (adjusting all the printf formats accordingly), and use the right hex values, instead.