Updated Driver.List after discrimination
Due to the request to keep also older Metasystem UPSs, we decided to modify a bit the Driver List (some models such as Megaline Series are Legrand and also Metasystem branded). we will keep the same driver for the 2 brands because we added in metasys.c a discrimination variable about battery SOC Data (The only command that was not available in older units).